+++ CHANGES.md
@@ -7,7 +7,7 @@ Next Release (5.5.0)
 
 Features
 --------
-* [#1131](https://github.com/java-native-access/jna/pull/1131): Add CoreFoundation, IOKit, and DiskArbitration mappings in `c.s.j.p.mac` with a `CoreFoundationUtil` class. [@dbwiddis](https://github.com/dbwiddis).
+* [#1131](https://github.com/java-native-access/jna/pull/1131): Add CoreFoundation, IOKit, and DiskArbitration mappings in `c.s.j.p.mac`. [@dbwiddis](https://github.com/dbwiddis).
 
 Bug Fixes
 ---------